requirements.txt
streamlit
langchain
langchain-google-genai
langgraph
google-generativeai
textstat
python-dotenv
